Update personal info display and add User Page function
The personal info display was restructured to be more concise, relocating the 'View Portfolio' button to the front and removing the 'Edit' button. A new function was added in views.py for a dedicated User Page, including the specification of the URL path for it. Resume logic was also modified, and templates have been updated in accordance.