Add 'Readme.txt' in 'scss' directory and new 'aos.js' and 'aos.css' files
The changes include the addition of new 'Readme.txt' guidance file in the 'scss' directory, explaining availability of .scss files in the pro version. This was accompanied by the introduction of two new files in 'vendor/aos' directory, 'aos.js' and 'aos.css', likely for handling new Animate On Scroll (AOS) functionalities.