Update job search form in home page
The job search form in the home page has been enhanced with new search options. This update will allow users to not only search for specific job listings but also search for users. The user interface of the form has been updated to accommodate these changes, for a more intuitive user experience.