Update website templates
The templates for user listings and home on the techlister website have been updated. For user listings, the code now includes a static file and an inner header, as well as a footer. The home template has been updated to change the placeholder text for the search input field, aiming to improve user experience.