Add restricted access page and improve user data privacy
A restricted access page was added to handle unauthorized access attempts. All parts of the site that previously just checked if a user was authenticated have been updated to also check if the request user is authorized to access or manage a particular resource. This enhancement significantly improves user data privacy throughout the site.