Implement read message status update
The code changes mainly address the read status update of messages. Now, when a user replies to a message, the main message's read status is set to False, showing that a new message has arrived. Similarly, when a user views a message, the read status of both the message and its replies are set to True. For visual indication on the user interface, a "View" button for read messages and a "Read" button for unread messages have been added