Update message viewing and fetching queries
The main change in the update is on how messages are fetched and viewed. Previously, the system would only display and fetch messages that a user is the recipient of. Now, it's adjusted to fetch and display both sent and received messages. Read and unread messages are distinguished using different buttons on the recruiter's profile page.