Update education form to use AJAX
Education form has been updated to use AJAX for form submission, and a new class has been added for identifying education rows for AJAX replacements. The edited education section of the user's profile will be updated without the need for a page refresh. Changes have been made directly in the templates via the usage of select attributes and the HttpResponse('') method in the view.