Update HTML templates for user education and experience
The variable name 'exp' was changed to 'experience' for clarity and consistency across the application. Also, changes were made to how elements are replaced in the DOM. The 'hx-swap' attribute was updated from 'outHTML' to 'outerHTML' and 'hx-target' and 'hx-include' values were likewise updated to target closer elements in the DOM tree.