Added UserSkill functionality to website
A UserSkillForm was created in forms.py and a corresponding template was generated in add_user_skill.html. Changes were made to users/templates/ai_resume.html, skill_list.html and views.py to accommodate the addition and deletion of UserSkills. The URL configuration in users/urls.py was also updated to route UserSkillForm submissions.