Enhanced AI resume page and its backend support
In the 'ai_resume.html' template, changes were made to improve the design and insert new features. Now, 'AI Resume', 'About Me', and 'Experience' sections are presented in a structured format with the help of card views. Backend support for these new sections is implemented in 'ai_resume' view and two new urls are added in 'users/urls.py' for the new functionalities.