Update QdrantClient initialization and modify search label
The QdrantClient now uses environment variables for initialization. Also, a slight modification was made to the search label.

Specifically, user-sensitive details such as API keys and URLs are now extracted directly from the environment, making the code more secure and easier to manage. In the template file, the label for user search has been changed from "Search users" to "Search For Talent" to better reflect the functionality.