Refactor user control section in user_listings.html
The div id "saved-user" is now only added if a user id exists. Resolved an issue where a div was being created even if the user id didn't exist, potentially leading to empty sections or unnecessary divs being added to the HTML. The order of control has been changed to improve page stability and layout.