Updated settings and modified .gitignore in techlister
The commit introduces a logging configuration to the settings.py file in the techlister directory. The logger writes to a specific file on debug level. Also, the list of allowed hosts is adjusted, modifying the entries for more specific ones. Additionally, the static URL and root have been changed. The settings.py file is also removed from the .gitignore list.