Add notice about free services on home and recruiter profile pages
A new notice has been added to the website's home page and recruiter profile page, informing users that our service is currently free of charge while we ensure the optimal functioning of our AI system. This addition aims to provide transparency to users about the ongoing development and refinement of the system.