Added 'active' field to PersonalInformation model
An 'active' field was introduced to the PersonalInformation model to keep track of users' active status. This involved not only updating the PersonalInformation model, but also making corresponding changes to the forms.py file and the SQL query in export_users.py. Additionally, the edit_personalinfo.html template was adjusted to include a checkbox for the 'active' status. The new field is set to 'true' by default.